+++ src/dom/removeNode.ts
@@ -1,9 +1,9 @@
-export default function removeNode(node: undefined | null): undefined;
-export default function removeNode<N extends Node>(node: N): N;
+function removeNode(node: undefined | null): undefined;
+function removeNode<N extends Node>(node: N): N;
 
-export default function removeNode<N extends Node>(
-	node: N | undefined | null
-): N | undefined {
+function removeNode<N extends Node>(node: N | undefined | null): N | undefined {
 	node && node.parentNode && node.parentNode.removeChild(node);
 	return node || undefined;
 }
+
+export default removeNode;