+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- fix: Nissing generic/fallback type signatures for `removeNode`, `curry`, `curryRight`
 
 ## 4.2.1